Some correction

+++ src/main/java/kz/sdu/chat/mainservice/constants/ValueConstants.java
@@ -8,5 +8,5 @@ import java.time.ZoneId;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class ValueConstants {
     public static final ZoneId ZONE_ID = ZoneId.of("UTC+00:00");
-    public static final Long ChatTokenLimit = 2L;
+    public static final Double ChatTokenLimit = 2.0;
 }
